refactor(structure): Reorganize project folder structure for improved scalability and maintainability
- Restructure project directories and file organization for better code discoverability
- Implement scalable folder hierarchy to support future development growth
- Reorganize source files into logical groupings for enhanced navigation
- Optimize package structure to improve code organization and maintainability
- Update file paths and import statements to reflect new organizational structure
