+++ src/main/java/module-info.java
@@ -12,4 +12,6 @@ module mx.edu.utez.warehousemanagerfx {
     exports mx.edu.utez.warehousemanagerfx;
     opens mx.edu.utez.warehousemanagerfx.models to javafx.fxml;
     exports mx.edu.utez.warehousemanagerfx.models;
+    exports mx.edu.utez.warehousemanagerfx.controllers;
+    opens mx.edu.utez.warehousemanagerfx.controllers to javafx.fxml;
 }
