--- src/main/java/mx/edu/utez/warehousemanagerfx/SuperAdminController.java
@@ -1,4 +1,0 @@
-package mx.edu.utez.warehousemanagerfx;
-
-public class SuperAdminController {
-}
