+++ src/main/java/mx/edu/utez/warehousemanagerfx/controllers/SuperAdminController.java
@@ -1,0 +1,4 @@
+package mx.edu.utez.warehousemanagerfx.controllers;
+
+public class SuperAdminController {
+}
