+++ src/main/java/mx/edu/utez/warehousemanagerfx/utils/routes/FXMLRoutes.java
@@ -1,0 +1,9 @@
+package mx.edu.utez.warehousemanagerfx.utils.routes;
+
+public class FXMLRoutes {
+    public static final String LOGIN = "/mx/edu/utez/warehousemanagerfx/fxml/login/LoginWindow.fxml";
+    public static final String SUPERADMIN = "/mx/edu/utez/warehousemanagerfx/fxml/superadmin/SuperAdminWindow.fxml";
+    public static final String ADMIN = "/mx/edu/utez/warehousemanagerfx/fxml/admin/AdminWindow.fxml";
+    public static final String WAREHOUSE_CARD = "/mx/edu/utez/warehousemanagerfx/fxml/warehouse/WarehouseCard.fxml";
+    public static final String WAREHOUSE_LIST = "/mx/edu/utez/warehousemanagerfx/fxml/warehouse/WarehouseList.fxml";
+}
