+++ src/main/java/mx/edu/utez/warehousemanagerfx/utils/routes/ImageRoutes.java
@@ -1,0 +1,6 @@
+package mx.edu.utez.warehousemanagerfx.utils.routes;
+
+public class ImageRoutes {
+    public static final String THUMBNAILS_BASE = "/mx/edu/utez/warehousemanagerfx/img/thumbnails/";
+    public static final String MAIN_APP_ICON = "/mx/edu/utez/warehousemanagerfx/img/logos/AppLogo.png";
+}
