+++ src/main/resources/mx/edu/utez/warehousemanagerfx/fxml/superadmin/SuperAdminWindow.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.shape.*?>
 <?import javafx.scene.text.*?>
 
-<AnchorPane maxHeight="500.0" maxWidth="600.0" minHeight="500.0" minWidth="600.0" prefHeight="500.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/17.0.12">
+<AnchorPane xmlns:fx="http://javafx.com/fxml/1" fx:controller="mx.edu.utez.warehousemanagerfx.controllers.SuperAdminController" maxHeight="500.0" maxWidth="600.0" minHeight="500.0" minWidth="600.0" prefHeight="500.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/17.0.12">
    <Rectangle arcHeight="5.0" arcWidth="5.0" fill="#00a8e8" height="75.0" stroke="TRANSPARENT" strokeType="INSIDE" width="600.0" />
    <Label layoutX="10.0" layoutY="17.0" text="¡Welcome Super Admin!" textFill="WHITE">
       <font>
@@ -15,7 +15,7 @@
    </Label>
    <Circle fill="WHITE" layoutX="554.0" layoutY="37.0" radius="32.0" stroke="TRANSPARENT" strokeType="INSIDE" />
    <ImageView fitHeight="64.0" fitWidth="64.0" layoutX="522.0" layoutY="5.0" pickOnBounds="true" preserveRatio="true">
-      <Image url="@img/icon/ProfileSA.png" />
+      <Image url="@../../img/avatars/ProfileSA.png" />
    </ImageView>
    <Button layoutX="415.0" layoutY="21.0" mnemonicParsing="false" prefHeight="31.0" prefWidth="79.0" text="Account">
       <font>
@@ -38,7 +38,7 @@
       </font>
    </Button>
    <ImageView fitHeight="150.0" fitWidth="200.0" layoutX="80.0" layoutY="124.0" pickOnBounds="true" preserveRatio="true">
-      <Image url="@img/icon/Admin.jpg" />
+      <Image url="@../../img/icons/Admin.jpg" />
    </ImageView>
    <Button layoutX="30.0" layoutY="398.0" mnemonicParsing="false" prefHeight="51.0" prefWidth="250.0" text="Adding a new Administrator">
       <font>
@@ -51,6 +51,6 @@
       </font>
    </Button>
    <ImageView fitHeight="150.0" fitWidth="200.0" layoutX="380.0" layoutY="124.0" pickOnBounds="true" preserveRatio="true">
-      <Image url="@img/icon/Branch.png" />
+      <Image url="@../../img/icons/Branch.png" />
    </ImageView>
 </AnchorPane>
