+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/avatars/Profile.png
