+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/avatars/ProfileSA.png
