+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/avatars/UserPicture.png
