+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/icons/ClearFilter.png
