+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/logos/UtezLogo.png
