+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/logos/WarehouseLogo.png
