+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/thumbnails/Warehouse1.jpg
