+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/thumbnails/Warehouse10.jpg
