+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/thumbnails/Warehouse11.jpg
