+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/thumbnails/Warehouse12.jpg
