+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/thumbnails/Warehouse2.jpg
