+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/thumbnails/Warehouse3.jpg
