+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/thumbnails/Warehouse4.jpg
