+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/thumbnails/Warehouse5.jpg
