+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/thumbnails/Warehouse6.jpg
