+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/thumbnails/Warehouse7.jpg
