+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/thumbnails/Warehouse8.jpg
