+++ src/main/resources/mx/edu/utez/warehousemanagerfx/img/thumbnails/Warehouse9.jpg
